+++ python/enumecg/__init__.py
@@ -6,7 +6,7 @@ The top level module provides the high level code generation API for the
 Enhanced Enum library.
 """
 
-__version__ = "0.2"
+__version__ = "0.3"
 __author__ = "Jaakko Moisio"
 
 from . import generators